+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -423,5 +423,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineEnableAutoUpdate");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCapacityReservation()
+        {
+            TestRunner.RunTestScript("Test-CapacityReservation");
+        }
     }
 }