+++ src/Compute/Compute/Manual/PSVirtualMachineScaleSetVMProfile.cs
@@ -36,5 +36,6 @@ namespace Microsoft.Azure.Commands.Compute.Automation.Models
         public string EvictionPolicy { get; set; }
         public BillingProfile BillingProfile { get; set; }
         public ScheduledEventsProfile ScheduledEventsProfile { get; set; }
+        public CapacityReservationProfile CapacityReservation { get; set; }
     }
 }