+++ src/Synapse/Synapse/Commands/ManagementCommands/WorkspaceActiveDirectoryAdministrator/SetAzureSynapseSqlActiveDirectoryAdministrator.cs
@@ -10,41 +10,62 @@ using System.Management.Automation;
 namespace Microsoft.Azure.Commands.Synapse
 {
     [Cmdlet(VerbsCommon.Set, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.Sql + SynapseConstants.ActiveDirectoryAdministrator,
-        DefaultParameterSetName = SetByNameParameterSet, SupportsShouldProcess = true)]
+        DefaultParameterSetName = SetByNameAndDisplayNameParameterSet, SupportsShouldProcess = true)]
     [OutputType(typeof(PSWorkspaceAadAdminInfo))]
     public class SetAzureSynapseSqlActiveDirectoryAdministrator : SynapseManagementCmdletBase
     {
-        private const string SetByNameParameterSet = "SetByNameParameterSet";
-        private const string SetByInputObjectParameterSet = "SetByInputObjectParameterSet";
-        private const string SetByResourceIdParameterSet = "SetByResourceIdParameterSet";
+        private const string SetByNameAndDisplayNameParameterSet = "SetByNameAndDisplayNameParameterSet";
+        private const string SetByInputObjectAndDisplayNameParameterSet = "SetByInputObjectAndDisplayNameParameterSet";
+        private const string SetByResourceIdAndDisplayNameParameterSet = "SetByResourceIdAndDisplayNameParameterSet";
+        private const string SetByNameAndObjectIdParameterSet = "SetByNameAndObjectIdParameterSet";
+        private const string SetByInputObjectAndObjectIdParameterSet = "SetByInputObjectAndObjectIdParameterSet";
+        private const string SetByResourceIdAndObjectIdParameterSet = "SetByResourceIdAndObjectIdParameterSet";
 
-        [Parameter(ParameterSetName = SetByNameParameterSet, Mandatory = false,
+        [Parameter(ParameterSetName = SetByNameAndDisplayNameParameterSet, Mandatory = false,
+            HelpMessage = HelpMessages.ResourceGroupName)]
+        [Parameter(ParameterSetName = SetByNameAndObjectIdParameterSet, Mandatory = false,
             HelpMessage = HelpMessages.ResourceGroupName)]
         [ResourceGroupCompleter()]
         [ValidateNotNullOrEmpty]
         public string ResourceGroupName { get; set; }
 
-        [Parameter(ParameterSetName = SetByNameParameterSet,
+        [Parameter(ParameterSetName = SetByNameAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.WorkspaceName)]
+        [Parameter(ParameterSetName = SetByNameAndObjectIdParameterSet,
             Mandatory = true, HelpMessage = HelpMessages.WorkspaceName)]
         [ResourceNameCompleter(ResourceTypes.Workspace, nameof(ResourceGroupName))]
         [ValidateNotNullOrEmpty]
         public string WorkspaceName { get; set; }
 
-        [Parameter(ValueFromPipeline = true, ParameterSetName = SetByInputObjectParameterSet,
+        [Parameter(ValueFromPipeline = true, ParameterSetName = SetByInputObjectAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.WorkspaceObject)]
+        [Parameter(ValueFromPipeline = true, ParameterSetName = SetByInputObjectAndObjectIdParameterSet,
             Mandatory = true, HelpMessage = HelpMessages.WorkspaceObject)]
         [ValidateNotNull]
         public PSSynapseWorkspace InputObject { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByResourceIdParameterSet,
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByResourceIdAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.WorkspaceResourceId)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByResourceIdAndObjectIdParameterSet,
             Mandatory = true, HelpMessage = HelpMessages.WorkspaceResourceId)]
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.DisplayName)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByNameAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.DisplayName)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByInputObjectAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.DisplayName)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByResourceIdAndDisplayNameParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.DisplayName)]
         [ValidateNotNullOrEmpty()]
         public string DisplayName { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = false, HelpMessage = HelpMessages.ObjectId)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByNameAndObjectIdParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.ObjectId)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByInputObjectAndObjectIdParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.ObjectId)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByResourceIdAndObjectIdParameterSet,
+            Mandatory = true, HelpMessage = HelpMessages.ObjectId)]
         [ValidateNotNullOrEmpty()]
         public Guid ObjectId { get; set; }