+++ src/Resources/Resources/ChangeLog.md
@@ -19,8 +19,12 @@
 -->
 
 ## Upcoming Release
+* Use JsonExtensions to serialize deserialize JSON objects to ensure the use of custom serialization settings [#15552]
+
+## Version 4.3.0
 * Fixed bug with `PSResource` where some constructors left `SubscriptionId` property unassigned/null.  [#10783]
 * Added support for creating and updating Template Spec in Bicep file [#15313]
+* Added `-ProceedIfNoChange` parameter to deployment create cmdlets.
 
 ## Version 4.2.0
 * Allowed naming the deployment when testing deployments [#11497]